+++ install.php
@@ -6,7 +6,7 @@ rex_sql_table::get(rex::getTable('pagestats_dump'))
     ->ensureColumn(new rex_sql_column('browsertype', 'varchar(255)'))
     ->ensureColumn(new rex_sql_column('brand', 'varchar(255)'))
     ->ensureColumn(new rex_sql_column('model', 'varchar(255)'))
-    ->ensureColumn(new rex_sql_column('url', 'varchar(255)'))
+    ->ensureColumn(new rex_sql_column('url', 'text'))
     ->ensureColumn(new rex_sql_column('date', 'date'))
     ->ensureColumn(new rex_sql_column('hour', 'int'))
     ->ensureColumn(new rex_sql_column('weekday', 'int'))
@@ -25,7 +25,7 @@ rex_sql_table::get(rex::getTable('pagestats_hash'))
     ->ensure();
 
 rex_sql_table::get(rex::getTable('pagestats_referer'))
-    ->ensureColumn(new rex_sql_column('referer', 'varchar(255)'))
+    ->ensureColumn(new rex_sql_column('referer', 'text'))
     ->ensureColumn(new rex_sql_column('count', 'int'))
     ->ensureColumn(new rex_sql_column('date', 'date'))
     ->ensure();