+++ Changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [unreleased] - xx.xx.xxxx
+## [1.0.0-rc.1] - 30.09.2021
 
 ### Added
 - permissions